+++ examples/java/Timer/Timer.java
@@ -1,4 +1,5 @@
-import org.openpatch.scratch.*;
+import org.openpatch.scratch.Sprite;
+import org.openpatch.scratch.Stage;
 
 public class Timer extends Stage {
   public Timer() {
@@ -12,7 +13,7 @@ public class Timer extends Stage {
 }
 
 class TimerSprite extends Sprite {
-  int x = 0;
+  int x = -1800;
 
   public TimerSprite() {
     super();
@@ -28,26 +29,26 @@ class TimerSprite extends Sprite {
   }
 
   public void run() {
-    int y = 20;
-    x += 20;
+    int y = -140;
+    this.x += 20;
     if (this.getTimer("every").everyMillis(600)) {
       this.getPen().down();
       this.getPen().setColor(20);
-      this.setPosition(x, y);
+      this.setPosition(this.x, y);
       this.getPen().up();
     }
     y += 40;
     if (this.getTimer("for").forMillis(600)) {
       this.getPen().down();
       this.getPen().setColor(60);
-      this.setPosition(x, y);
+      this.setPosition(this.x, y);
       this.getPen().up();
     }
     y += 40;
     if (this.getTimer("after").afterMillis(600)) {
       this.getPen().down();
       this.getPen().setColor(100);
-      this.setPosition(x, y);
+      this.setPosition(this.x, y);
       this.getPen().up();
     }
     y += 40;
@@ -55,28 +56,28 @@ class TimerSprite extends Sprite {
       this.getPen().down();
 
       this.getPen().setColor(140);
-      this.setPosition(x, y);
+      this.setPosition(this.x, y);
       this.getPen().up();
     }
     y += 40;
     if (this.getTimer("interval2").intervalMillis(600, true)) {
       this.getPen().down();
       this.getPen().setColor(180);
-      this.setPosition(x, y);
+      this.setPosition(this.x, y);
       this.getPen().up();
     }
     y += 40;
     if (this.getTimer("interval3").intervalMillis(600, 300)) {
       this.getPen().down();
       this.getPen().setColor(220);
-      this.setPosition(x, y);
+      this.setPosition(this.x, y);
       this.getPen().up();
     }
     y += 40;
     if (this.getTimer("interval4").intervalMillis(600, 300, true)) {
       this.getPen().down();
       this.getPen().setColor(255);
-      this.setPosition(x, y);
+      this.setPosition(this.x, y);
       this.getPen().up();
     }
   }