+++ examples/reference/PenSetSize.java
@@ -1,6 +1,7 @@
-import org.openpatch.scratch.*;
-import org.openpatch.scratch.extensions.pen.*;
-import org.openpatch.scratch.extensions.recorder.*;
+import org.openpatch.scratch.Stage;
+import org.openpatch.scratch.Window;
+import org.openpatch.scratch.extensions.pen.Pen;
+import org.openpatch.scratch.extensions.recorder.GifRecorder;
 
 public class PenSetSize {
   public PenSetSize() {
@@ -15,7 +16,7 @@ public class PenSetSize {
       if (myStage.getTimer().everyMillis(500)) {
         myPen.setSize(myStage.pickRandom(5, 20));
       }
-      myPen.setPosition(myStage.pickRandom(0, myStage.getWidth()), 45);
+      myPen.goToRandomPosition();
       myPen.up();
       myStage.wait(200);
     }